Roulette data created
